Bill S9725 version original - William Weber
BILL NUMBER: S9725 SPONSOR: WEBER TITLE OF BILL: An act to amend the tax law, in relation to authorizing the town of Clarkstown to impose a hotel and motel tax; and providing for the repeal of such provisions upon expiration thereof PURPOSE: This bill will allow the Town of Clarkstown to adopt and amend local laws to impose a tax on hotel or motel rooms within the Town of Clarks- town. SUMMARY OF PROVISIONS: Section 1.