Bill S9212A version A - Pete Harckham
BILL NUMBER: S9212A SPONSOR: HARCKHAM TITLE OF BILL: An act authorizing the town of Yorktown, county of Westchester, to alienate and discontinue the use of certain parklands PURPOSE OR GENERAL IDEA OF BILL: Authorizes the town of Yorktown, county of Westchester, to alienate and discontinue the use of certain parklands. SUMMARY OF PROVISIONS: Section 1 - Authorizes Yorktown to discontinue the use as parkland and alienate the lands described in the bill.