Bill S643E version E - Brian Kavanagh
BILL NUMBER: S643E SPONSOR: KAVANAGH TITLE OF BILL: An act to amend the environmental conservation law, in relation to rechargeable battery recycling PURPOSE: To maximize the removal of unwanted rechargeable batteries from the solid waste stream. SUMMARY OF PROVISIONS: Section one of the bill amends the heading of article 27 of the environ- mental conservation law. Section two of the bill amends the title heading of title 18 of article 27 of the environmental conservation law.