Bill S9735 version original - Jamaal Bailey
BILL NUMBER: S9735 SPONSOR: BAILEY TITLE OF BILL: An act to amend chapter 101 of the laws of 2014 amending the vehicle and traffic law, the general municipal law, and the public officers law relating to owner liability for failure of operator to comply with traffic-control indications in the city of Mt. Vernon, in relation to extending the provisions of such chapter PURPOSE: To prevent the deaths and injuries caused by motorists running red lights, by authorizing the City of Mt.