Bill S9729 version original - Brad Hoylman-Sigal
BILL NUMBER: S9729 SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the criminal procedure law, in relation to the examina- tion of certain sealed records PURPOSE:: This bill would expressly provide for certain sealed records to be made available to OSC's officers and employees for purposes of conducting a duly authorized examination into the financial affairs of town and village justice courts. SUMMARY OF PROVISIONS:: Section 1 of this bill amends Criminal Procedure Law ("CPL") § 160.