Bill S9738A version original - Neil D. Breslin
BILL NUMBER: S9738 SPONSOR: BRESLIN TITLE OF BILL: An act in relation to authorizing the lease of certain lands for the Albany NanoTech Complex PURPOSE OR GENERAL IDEA OF BILL: Authorizes the State University of New York at Albany to lease certain lands to the Fuller Road Management Corporation for purposes of the Albany NanoTech Complex. SUMMARY OF PROVISIONS: Section one declares the legislative findings.