Bill S9740 version original - William Weber
BILL NUMBER: S9740 SPONSOR: WEBER TITLE OF BILL: An act to amend the tax law, in relation to authorizing the town of Ramapo to impose a hotel and motel tax; and providing for the repeal of such provisions upon expiration thereof PURPOSE: This bill will allow the Town of Ramapo to adopt and amend local laws to impose a tax on hotel or motel rooms within the Town of Ramapo. SUMMARY OF PROVISIONS: Section 1.