Bill S9743 version original - José M. Serrano
BILL NUMBER: S9743 SPONSOR: SERRANO TITLE OF BILL: An act to amend the state finance law, in relation to the New York state music grant fund Purpose of Bill: An act to amend the finance law, in relation to the New York State coun- cil on the arts. SUMMARY OF PROVISIONS: The proposed bill amends Section 97-v of the finance law to allow unused funds to be designated for the not-for-profit musical entities across the state. See 6 below. PROPOSED EFFECTIVE DATE: Immediate.