Bill S9436A version A - Jeremy Cooney
BILL NUMBER: S9436A SPONSOR: COONEY TITLE OF BILL: An act directing the commissioner of motor vehicles to conduct a study of official inspection stations, safety inspections and fees PURPOSE OR GENERAL IDEA OF BILL: Requires the NYS Department of Motor Vehicles to conduct a study on the appropriate fees and subjects of the annual motor vehicle safety and emissions inspection. SUMMARY OF PROVISIONS: Section 1: Directs the NYS Department of Motor Vehicles.