Bill S9750 version original - Michael Gianaris
BILL NUMBER: S9750 SPONSOR: GIANARIS TITLE OF BILL: An act to amend the penal law, in relation to restrictions on the receipt of ammunition PURPOSE: To amend the Penal Law to require ammunition background checks to be denied if the prospective purchaser cannot lawfully possess a rifle or shotgun under federal or New York State law. SUMMARY OF PROVISIONS: Section 1 of this bill would amend Penal Law § 400.