Bill S9756 version original - Lea Webb
BILL NUMBER: S9756 SPONSOR: WEBB TITLE OF BILL: An act to amend chapter 152 of the laws of 2020 amending the environ- mental conservation law relating to authorizing hunting big game by rifle in the county of Tompkins, in relation to extending the effective- ness thereof PURPOSE: Extends provisions authorizing hunting big game by rifle in the county of Tompkins SUMMARY OF PROVISIONS: Section one of the bill provides for a two-year extender. Section two is the effective date.