Bill S3036A version A - Samra Brouk
BILL NUMBER: S3036A SPONSOR: BROUK TITLE OF BILL: An act to amend the education law and the social services law, in relation to the practice of a licensed school psychology practitioner PURPOSE: This bill provides for licensure of school psychology practitioners and defines the educational and experiential qualifications necessary for licensure and continued practice. SUMMARY OF SPECIFIC PROVISIONS: Section one provides the legislative intent.