Bill S1400A version A - John Liu
BILL NUMBER: S1400A SPONSOR: LIU TITLE OF BILL: An act to amend the environmental conservation law, in relation to false claims about recyclability and plastic container labelling PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to ensure that plastic packaging is labeled correctly. SUMMARY OF PROVISIONS: Section 1 states the legislative findings and declaration.