Bill S5465A version A - Robert Jackson
BILL NUMBER: S5465A SPONSOR: JACKSON TITLE OF BILL: An act to amend the executive law, in relation to prohibiting employers from discriminating against individuals based on such individual's status as a caregiver PURPOSE: The purpose of the bill is to provide protections for caregivers under the Human Rights Law. SUMMARY OF PROVISIONS: Section 1 amends the definition of "reasonable accommodation" in Section 292 of the Executive Law to address caregivers' needs.