Bill S9753 version original - Thomas F. O'Mara
BILL NUMBER: S9753 SPONSOR: O'MARA TITLE OF BILL: An act to amend chapter 545 of the laws of 2011, amending the real prop- erty tax law, relating to establishing a residential-commercial exemption program in certain counties, in relation to the effectiveness thereof PURPOSE: This bill extends the residential-commercial exemption program in Livingston and Steuben County, New York. SUMMARY OF PROVISIONS: Section 1.