Bill S9504A version A - Jeremy Cooney
BILL NUMBER: S9504A SPONSOR: COONEY TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to increasing penalties for violation of overtaking and passing a school bus PURPOSE OR GENERAL IDEA OF BILL: To increase the penalties for passing a stopped school bus. SUMMARY OF PROVISIONS: Section 1: Amends Subdivision c of section 1174 of the vehicle and traf- fic law to allow for additional penalties for an individual convicted of passing a stopped school bus.