Bill S9754 version original - Iwen Chu
BILL NUMBER: S9754 SPONSOR: CHU TITLE OF BILL: An act to amend the executive law, the public authorities law and the county law, in relation to requiring certain documents and forms to be provided in multiple languages PURPOSE OR GENERAL IDEA OF BILL: The purpose of the bill is to expand the New York state language access policy. SUMMARY OF PROVISIONS: Section 1. Provide for the translation of essential public documents such as forms and instructions, and include relevant data sources.