Bill S9755 version original - Monica Martinez
BILL NUMBER: S9755 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the real property tax law, in relation to establishing a real property tax exemption for widows and widowers of police officers killed in the line of duty PURPOSE: The purpose of this legislation is to establish a real property tax exemption for widows and widowers of police officers killed in the line of duty. SUMMARY OF PROVISIONS: Section 1.