Bill S9747 version original - Jessica Ramos
BILL NUMBER: S9747 SPONSOR: RAMOS TITLE OF BILL: An act authorizing the city of New York to discontinue certain parkland in Flushing Meadows Corona Park in the borough of Queens currently consisting of asphalt parking lots to allow for a new development with at least fifty acres of parkland PURPOSE: Authorize the city of New York to discontinue certain parkland in Flush- ing Meadows Corona Park in the borough of Queens for the purposes described. SUMMARY OF PROVISIONS: Section 1.