Bill S9746 version original - Lea Webb
BILL NUMBER: S9746 SPONSOR: WEBB TITLE OF BILL: An act to amend the domestic relations law, in relation to automatic orders in matrimonial actions 2. SOURCE OF BILL This bill is being introduced at the request of the Judiciary, upon the recommendation of the Chief Administrative Judge's Matrimonial Practice Advisory and Rules Committee. 3.