Bill S158E version E - Liz Krueger
BILL NUMBER: S158E SPONSOR: KRUEGER TITLE OF BILL: An act to amend the general business law, in relation to providing for the protection of health information PURPOSE OR GENERAL IDEA OF BILL: This bill would govern companies that collect and sell healthcare infor- mation and provides additional rights and protections to users related to the sale and of their private health information. SUMMARY OF SPECIFIC PROVISIONS: Section one amends the general business law by adding a new article 42.