Bill S8485B version B - Brad Hoylman-Sigal
BILL NUMBER: S8485B SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the estates, powers and trusts law, in relation to the payment and distribution of damages in wrongful death actions SUMMARY OF PROVISIONS OF BILL: Section 1 amends EPTL section 5-4.1 to extend the time permitted to bring a wrongful death action by one year. Section 2 amends EPTL section 5-4.3, to permit recovery of damages for grief and anguish when a tortfeasor is found liable for causing a death.