Bill S9762 version original - James Tedisco
BILL NUMBER: S9762 SPONSOR: TEDISCO TITLE OF BILL: An act to amend the criminal procedure law and the executive law, in relation to enacting the "child survivor privacy act" PURPOSE OR GENERAL IDEA OF BILL: To provide greater privacy protection for minors who are victims or witnesses and subject to providing testimony. SUMMARY OF PROVISIONS: § 1. This act shall be known and may be cited as the "child survivor privacy act". § 2.