Bill S9775A version original - Jeremy Cooney
BILL NUMBER: S9775 SPONSOR: COONEY TITLE OF BILL: An act to amend chapter 698 of the laws of 1991 amending the public authorities law and the education law, relating to the construction and financing of facilities by the dormitory authority for private schools approved pursuant to article 89 of the education law, in relation to including Mary Cariola Children's Center Inc. and Mary Cariola Founda- tion Inc.