Bill S4877A version A - Shelley Mayer
BILL NUMBER: S4877A SPONSOR: MAYER TITLE OF BILL: An act to amend the state finance law, in relation to state agency contracts with not-for-profit corporations PURPOSE: To require agencies with significantly delayed renewal contracts to report on the status of delayed contracts and obstacles to contracts being executed in a timely manner. SUMMARY OF PROVISIONS: Section one of the bill amends section 179-q of the state finance law to define "noncompliant state agency.