Bill S2158A version A - Anthony H. Palumbo
BILL NUMBER: S2158A SPONSOR: PALUMBO TITLE OF BILL: An act in relation to the eligibility of enrolled members of Southampton Village Ocean Rescue for a tax exemption on real property PURPOSE: This legislation recognizes the work and commitment of the Southampton Village Ocean Rescue by granting eligibility for tax exemption on real property tax to enrolled members. SUMMARY OF PROVISIONS: Section 1.