Bill S9771 version original - Jessica Scarcella-Spanton
BILL NUMBER: S9771 SPONSOR: SCARCELLA-SPANTON TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to special license plates for vehicles owned or used by people with disabilities SUMMARY OF PROVISIONS: Section 1. Allows for the parents or guardians of a person with a disa- bility to register for distinctive license plates that reflect the status of an occupant of that vehicle being a person with a disability. Section 2. This shall take effect immediately.