Bill S9763 version original - Shelley Mayer
BILL NUMBER: S9763 SPONSOR: MAYER TITLE OF BILL: An act to amend the village law, in relation to conducting referenda in villages with November elections PURPOSE: To ensure that referenda are conducted in accordance with the political calendar prescribed by the election law. SUMMARY OF PROVISIONS: Section 1 of the bill amends C 9-900 of the village law, by adding a new subdivision four.