Bill S8790A version A - Alexis Weik
BILL NUMBER: S8790A SPONSOR: WEIK TITLE OF BILL: An act to amend the highway law, in relation to designating a portion of the state highway system as "Charlie Bunger Sr. Way" PURPOSE OR GENERAL IDEA OF BILL: Designates a portion of the state highway system as "Charlie Bunger Sr. SUMMARY OF SPECIFIC PROVISIONS: Section 1. The highway law is amended by adding a new section 344-p to designate a portion of Montauk Highway in the Village of Babylon as "Charlie Bunger Sr. Way".