Bill S9741A version A - Anthony H. Palumbo
BILL NUMBER: S9741A SPONSOR: PALUMBO TITLE OF BILL: An act authorizing the town of East Hampton to alienate and convey certain parcels of land used as parkland for the purpose of a Suffolk County roadway improvement project PURPOSE: Authorizes the town of East Hampton to alienate and convey certain parcels of land used as parkland for the purpose of a Suffolk County roadway improvement project. SUMMARY OF PROVISIONS: Section 1.