Bill S9773 version original - Pete Harckham
BILL NUMBER: S9773 SPONSOR: HARCKHAM TITLE OF BILL: An act to amend the tax law, in relation to authorizing an occupancy tax in the town of Ossining PURPOSE OR GENERAL IDEA OF BILL: The purpose of this legislation is to allow the Town of Ossining, West- chester County, to levy a hotel occupancy tax not to exceed three percent. SUMMARY OF PROVISIONS: Section 1.