Bill S2159A version A - Anthony H. Palumbo
BILL NUMBER: S2159A SPONSOR: PALUMBO TITLE OF BILL: An act in relation to the eligibility of enrolled members of the East Hampton Volunteer Ocean Rescue and Auxiliary Squad for a tax exemption on real property PURPOSE: This legislation recognizes the work and commitment of the East Hampton Volunteer Ocean Rescue and Auxiliary Squad by granting eligibility for tax exemption on real property tax to enrolled members. SUMMARY OF PROVISIONS: Section 1.