Bill S6419E version E - Brian Kavanagh
BILL NUMBER: S6419E SPONSOR: KAVANAGH TITLE OF BILL: An act to amend the environmental conservation law, in relation to establishing a mattress collection program PURPOSE OF BILL: This bill would require extended producer responsibility for the collection, reuse, recycling, or proper disposal of mattresses. SUMMARY OF PROVISIONS: Section one of the bill amends the environmental conservation law to add a new title 34 that would establish the mattress collection program.