Bill S9784 version original - Brad Hoylman-Sigal
BILL NUMBER: S9784 SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the domestic relations law, in relation to modifying the provisions regarding notice of a matrimonial action where the complaint is not personally served with the summons 2. SOURCE OF BILL This bill is being introduced at the request of the Judiciary, on the recommendation of the Chief Administrative Judge's Matrimonial Practice Advisory and Rules Committee. 3.