Bill S9779 version original - Brian Kavanagh
BILL NUMBER: S9779 SPONSOR: KAVANAGH TITLE OF BILL: An act to amend the environmental conservation law, in relation to extended producer responsibility for carpet PURPOSE OR GENERAL IDEA OF BILL: To provide antitrust provisions in the Carpet Collection Program. SUMMARY OF PROVISIONS: Section one of the bill amends the title heading of title 33 of article 27 of the environmental conservation law to read as "Extended Producer Responsibility for Carpet".