Bill S9221A version A - Anthony H. Palumbo
BILL NUMBER: S9221A SPONSOR: PALUMBO TITLE OF BILL: An act in relation to authorizing the town of Brookhaven to discontinue a portion of real property in the county of Suffolk as parkland and to grant easements to West Ferry Office, LLC for the construction of sewer facilities PURPOSE: To allow alienation of parkland in the town of Brookhaven for the purposes of connecting a sewer line at the Port Jefferson Ferry Termi- nal. SUMMARY OF PROVISIONS: Section 1.