Bill S931A version A - Roxanne J. Persaud
BILL NUMBER: S931A SPONSOR: PERSAUD TITLE OF BILL: An act to amend the public health law, in relation to prohibiting virginity examinations PURPOSE OR GENERAL IDEA OF BILL: The purpose of the bill is to prevent the performance of examinations on women that purport. to ascertain whether a woman has previously had vaginal intercourse. SUMMARY OF SPECIFIC PROVISIONS: Section 1 of the bill adds a new section 267 a to the public health law to forbid the performance of "virginity examinations".