Bill S9815 version original - James Skoufis
BILL NUMBER: S9815 SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the executive law, in relation to adding Moodna Creek in Orange County to the definition of inland waterways PURPOSE OR GENERAL IDEA OF BILL: Adds Moodna Creek in Orange County to the definition of inland waterways SUMMARY OF SPECIFIC PROVISIONS: Section 1: Amends Paragraph (c) of subdivision 4 of section 911 of the executive law to include Moodna Creek. Section 2: Sets the effective date.