Bill S9809 version original - Kevin Thomas
BILL NUMBER: S9809 SPONSOR: THOMAS TITLE OF BILL: An act authorizing the town of Hempstead to transfer ownership of certain parkland to the village of Freeport PURPOSE OR GENERAL IDEA OF BILL: This bill authorizes Town of Hempstead to convey Guy Lombardo Marina, located in the Village of Freeport, to the Village. SUMMARY OF PROVISIONS: Section 1 authorizes the Town of Hempstead to convey Guy Lombardo Marina to the Village of Freeport for use as a public park.