Bill S9806 version original - Joseph P. Addabbo Jr.
BILL NUMBER: S9806 SPONSOR: ADDABBO TITLE OF BILL: An act to amend the racing, pari-mutuel wagering and breeding law, in relation to improvement of the self-exclusion provisions PURPOSE OF THE BILL: This bill enhances the Gaming Commission's self-exclusion program. Current law provides for self-exclusion, but authorizes forfeiture of winnings in violation of self-exclusion only for prohibited commercial casino gambling.