Bill S8908A version A - Anthony H. Palumbo
BILL NUMBER: S8908A SPONSOR: PALUMBO TITLE OF BILL: An act relating to permitting the Setauket fire district to enter into contracts with the state university of New York at Stony Brook for the sale or donation of emergency equipment PURPOSE: To allow Setauket Fire District to store a surplus truck on the campus of Stony Brook University. SUMMARY OF PROVISIONS: Section 1.