Bill S3249B version A - Jamaal Bailey
BILL NUMBER: S3249A SPONSOR: BAILEY TITLE OF BILL: An act to amend the executive law, in relation to requiring the instal- lation of smoke detectors in common places of certain dwellings PURPOSE: The purpose of this bill is to make residential buildings safe from fire hazards. SUMMARY OF PROVISIONS: Section 1.