Bill S9836 version original - Robert Jackson
BILL NUMBER: S9836 SPONSOR: JACKSON TITLE OF BILL: An act to amend the state finance law, in relation to requiring a quar- terly report on the opioid stewardship fund PURPOSE: To ensure that Opioid Stewardship Funds are being utilized in an expedi- ent manner to combat the substance use disorder crisis. SUMMARY OF PROVISIONS: Section 1. Section 97-aaaaa of the state finance law is amended by adding a new subdivision 8 to read as follows: 8.