Bill S9332A version A - Jack M. Martins
BILL NUMBER: S9332A SPONSOR: MARTINS TITLE OF BILL: An act in relation to authorizing the town of Oyster Bay to discontinue as parklands and convey such parkland to the department of environmental conservation PURPOSE OR GENERAL IDEA OF BILL: Authorizes the town of Oyster Bay to discontinue as parklands and convey such parkland to the DEC SUMMARY OF PROVISIONS: Section 1: The Town.of Oyster Bay, upon consent, is hereby authorized to discontinue and. convey. town parkland.