Bill S9837 version original - Zellnor Myrie
BILL NUMBER: S9837 SPONSOR: MYRIE TITLE OF BILL: An act to amend the election law, in relation to curing ballots PURPOSE OR GENERAL IDEA OF BILL: To ensure that voters are able to cure ballot defects within the allowed timeframe. SUMMARY OF SPECIFIC PROVISIONS: Section 1 amends subdivision 3 of § 9-209 requiring the State Board of Elections to provide a postage paid envelope within which a voter may return a cure affirmation form.