Bill S9839 version original - Brad Hoylman-Sigal
BILL NUMBER: S9839 SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the family court act and the New York city civil court act, in relation to increasing the number of judges in certain courts PURPOSE: This bill increases the number of Family Court judges throughout the state as well as increases the number of judges of the Civil Court of the City of New York. SUMMARY OF PROVISIONS: Sections 1 and 3 amends the Family Court Act. Section 2 amends the New York city civil court act.