Bill S8608A version A - Jessica Ramos
BILL NUMBER: S8608A SPONSOR: RAMOS TITLE OF BILL: An act to amend the labor law, in relation to requiring contractors and subcontractors employed by the state to submit their payrolls or tran- scripts to the fiscal officer PURPOSE OR GENERAL IDEA OF BILL: To add a new section 220-j to Article 8 of the labor law to create an electronic certified payroll system applicable to all prevailing wage projects. SUMMARY OF PROVISIONS: Section 1.