Bill S6110C version C - George Borrello
BILL NUMBER: S6110C SPONSOR: BORRELLO TITLE OF BILL: An act to amend the public officers law and the village law, in relation to waiving the local residency requirement for the village clerk and village treasurer in the village of Portville PURPOSE: Waives the local residency requirement for public officers in the village of Portville. SUMMARY OF PROVISIONS: Section 1.