Bill S9833 version original - Shelley Mayer
BILL NUMBER: S9833 SPONSOR: MAYER TITLE OF BILL: An act in relation to apportioning funds to the East Ramapo central school district for services and expenses PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to provide East Ramapo Central School District with a spin up with certain limitations. SUMMARY OF PROVISIONS: Section 1 provides the legislative intent. Section 2 provides a spin up of $20M to East Ramapo Central School District.