Bill S5643A version A - Cordell Cleare
BILL NUMBER: S5643A SPONSOR: CLEARE TITLE OF BILL: An act to amend the general business law, in relation to prohibiting the sale of infant walkers and restricting the use of such infant walkers in certain settings PURPOSE OR GENERAL IDEA OF BILL: To ban the sale of infant walkers and restrict their use in certain settings. SUMMARY OF SPECIFIC PROVISIONS: Section 1 of the bill amends the general business law by adding a new section 399-k.