Bill S8907A version A - Kristen Gonzalez
BILL NUMBER: S8907A SPONSOR: GONZALEZ TITLE OF BILL: An act to direct the department of health complete a report on the impact of hospital closures on healthcare access in New York state PURPOSE: The bill requires the Department of Health to collect and evaluate detailed data on healthcare facilities throughout New York State. This data will be used to identify and address the specific needs of distressed healthcare facilities, thereby safeguarding healthcare access statewide.