Bill S3037A version A - Mark Walczyk
BILL NUMBER: S3037A SPONSOR: WALCZYK TITLE OF BILL: An act to amend the highway law, in relation to designating a portion of the state highway system as the "War Dog Memorial Highway" PURPOSE OR GENERAL IDEA OF BILL: Creates the "War Dog Memorial Highway" SUMMARY OF PROVISIONS: Section 1. The highway law is amended by adding a new section 344-p to read as follows: § 343-p.