Bill S9860 version original - Neil D. Breslin
BILL NUMBER: S9860 SPONSOR: BRESLIN TITLE OF BILL: An act to amend the insurance law, in relation to invoking coverage of a health insurance plan for long term care for an incapacitated individual PURPOSE: Grant authority to invoke coverage of a health insurance plan for long term care to an insured's legal spouse if not legally separated from the insured, or domestic partner. SUMMARY OF PROVISIONS: Section 1.