Bill S3249B version B - Jamaal Bailey
BILL NUMBER: S3249B SPONSOR: BAILEY TITLE OF BILL: An act to amend the executive law, the multiple dwelling law, the multi- ple residence law and the administrative code of the city of New York, in relation to requiring the installation of smoke detectors in common places of certain dwellings PURPOSE: The purpose of this bill is to make residential buildings safe from fire hazards. SUMMARY OF PROVISIONS: Section 1.