Bill S9856 version original - Brad Hoylman-Sigal
BILL NUMBER: S9856 SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the civil practice law and rules, in relation to alter- nate jurors This is one in a series of measures being introduced at the request of the Acting Chief Administrative Judge upon the recommendation of her Advisory Committee on Civil Practice. This measure would amend CPLR 4106 in relation to use of alternate jurors.