Bill S9866 version original - Luis R. Sepúlveda
BILL NUMBER: S9866 SPONSOR: SEPULVEDA TITLE OF BILL: An act to amend the education law, in relation to tuition benefit incen- tives for members of the New York state organized militia PURPOSE: This bill extends the recruitment incentive program to active military members for the attainment of a master's degree as well as a baccalau- reate degree. SUMMARY OF PROVISIONS: Section 1. This act shall be known as the "Technical Sergeant Joseph G. Lemm tuition benefit act". Section 2.