Bill S9870 version original - Monica Martinez
BILL NUMBER: S9870 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the retirement and social security law, in relation to an optional twenty-five year retirement plan for fire marshals I and supervisors of such title in Suffolk county PURPOSE: The purpose of this bill is to authorize an optional twenty-five year retirement plan for fire marshals in Suffolk County. SUMMARY OF PROVISIONS: Section 1.