Bill S9873 version original - Roxanne J. Persaud
BILL NUMBER: S9873 SPONSOR: PERSAUD TITLE OF BILL: An act to amend the penal law, in relation to the sale of ammunition for assault weapons PURPOSE OR GENERAL IDEA OF BILL: This bill would limit the sale of ammunition designed for use in assault weapons to twice the capacity of such weapon(s) to those authorized to possess said weapon(s) within a one hundred twenty-day period. SUMMARY OF SPECIFIC PROVISIONS: New York Penal Law Section 270.