Bill S9876 version original - Mario Mattera
BILL NUMBER: S9876 SPONSOR: MATTERA TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to establishing a penalty for interfering with the repossession of certain vehicles PURPOSE OR GENERAL IDEA OF BILL: To amend the vehicle and traffic law, in relation to establishing a penalty for interfering with the repossession of certain vehicles. SUMMARY OF PROVISIONS: Section 1.