Bill S9883 version original - George Borrello
BILL NUMBER: S9883 SPONSOR: BORRELLO TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to penalties for traffic infractions related to the operation of vehicles when approaching authorized emergency vehicles PURPOSE: This bill relates to penalties for traffic infractions related to the operation of vehicles when approaching authorized emergency vehicles. SUMMARY OF PROVISIONS: Section 1: Is amended by adding a new subdivision (e-1).