Bill S9887 version original - William Weber
BILL NUMBER: S9887 SPONSOR: WEBER TITLE OF BILL: An act in relation to apportioning funds to the East Ramapo central school district for services and expenses PURPOSE: The purpose of this bill is to provide the East Ramapo Central School District with a spin up. SUMMARY OF PROVISIONS: Section 1. States the legislative intent. Section 2. Provides a spin up of $20 million to the East Ramapo Central School District. Section 3. Sets the effective date.