Bill S9885A version original - Jake Ashby
BILL NUMBER: S9885 SPONSOR: ASHBY TITLE OF BILL: An act to amend the tax law, in relation to payment of a supplemental empire state child tax credit PURPOSE: To provide tax relief to those residents with a newborn. SUMMARY OF PROVISIONS: Section 1 establishes a refundable Supplemental Empire State Child Cred- it equal to $1,000 per newborn. Section 2 of the bill states this act shall take effect April 1, 2025. JUSTIFICATION: The costs to live in NYS are staggering.