Bill S9888 version original - Jake Ashby
BILL NUMBER: S9888 SPONSOR: ASHBY TITLE OF BILL: An act to amend the tax law, in relation to creating an in vitro ferti- lization treatment tax credit PURPOSE: To provide tax relief to those residents undergoing in vitro fertiliza- tion treatments. SUMMARY OF PROVISIONS: Section 1 establishes a refundable tax credit equal to 75% of the expenses for up to 3 cycles of in vitro fertilization treatments an d which are not covered by insurance. The amount shall not exceed $10,000.