Bill S9895 version original - Joseph A. Griffo
BILL NUMBER: S9895 SPONSOR: GRIFFO TITLE OF BILL: An act to amend the tax law, in relation to enacting the "grid resilien- cy act" PURPOSE: To prevent New York State from closing any power generation facility before new facilities are operational and provide tax credits to homeowners for the purchase of backup power systems in the event of an outage. SUMMARY OF PROVISIONS: 1.