Bill S9899 version original - Robert Ortt
BILL NUMBER: S9899 SPONSOR: ORTT TITLE OF BILL: An act to amend the environmental conservation law, in relation to prohibiting certain restrictions on motor vehicle sales PURPOSE: To prohibit certain restrictions on motor vehicle sales. SUMMARY OF PROVISIONS: Section 1: No agency, department, or division of state government, or a municipality will regulate the use or sale of motor vehicles by the public based on the energy or fuel used.