Bill S9909 version original - Jessica Scarcella-Spanton
BILL NUMBER: S9909 SPONSOR: SCARCELLA-SPANTON TITLE OF BILL: An act to amend the public authorities law, in relation to limiting the amount of the fine to be paid for violating the E-ZPass system SUMMARY OF PROVISIONS: Section 1. Amends public authorities' law by adding a new section 354-a by adding a violation cap and further establishes that if any person fails to pay a toll as required by the E-ZPass system, the authority shall not issue a fine exceeding the amount of the unpaid toll.