Bill S8993A version A - Jessica Ramos
BILL NUMBER: S8993A SPONSOR: RAMOS TITLE OF BILL: An act to amend the private housing finance law, in relation to estab- lishing a jobs and housing pilot program to create jobs in the construction industry and address the housing crisis; and making an appropriation therefor PURPOSE: To establish the Jobs and Housing Act and create a pilot program for building affordable housing. SUMMARY OF PROVISIONS: Section 1. This act shall be known and may be cited as the "jobs and housing act". Section 2.