Bill S9915 version original - Andrew Gounardes
BILL NUMBER: S9915 SPONSOR: GOUNARDES TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to injury of domestic and companion animals by motorists PURPOSE OR GENERAL IDEA OF BILL: To increase the penalties for reckless driving that results in the inju- ry or loss of a pet. SUMMARY OF PROVISIONS: Section one amends Section 1146 of the Vehicle and Traffic Law to add the term "companion animal" to the criteria of victims that are included under this section.