Bill S9923 version original - Jeremy Cooney
BILL NUMBER: S9923 SPONSOR: COONEY TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to the defi- nition of drugs SUMMARY OF PROVISIONS: Section 1. Amends section 114-a of the vehicle and traffic law to expand the definition of drug to any substance or combination of substances that impair, to any extent, physical or mental abilities. Section 2. Sets the effective date.