Bill S9924A version A - Iwen Chu
BILL NUMBER: S9924A SPONSOR: CHU TITLE OF BILL: An act to amend the penal law, in relation to assault on a food delivery person PURPOSE: The purpose of this legislation is to establish the Class E felony of assault on a food delivery worker. SUMMARY OF PROVISIONS: Section one amends the penal law by adding new section 120.19-a, which establishes the crime of assault on a food delivery person as a class E Felony.