Bill S9933 version original - Jeremy Cooney
BILL NUMBER: S9933 SPONSOR: COONEY TITLE OF BILL: An act to amend the education law, in relation to establishing Diwali as a school holiday SUMMARY OF SPECIFIC PROVISION: Section one amends subdivision 8 of section 3604 of the education law to add Diwali as a school holiday for all public schools. Section two establishes an effective date. JUSTIFICATION: Diwali, also known as Deepavali, is the Hindu festival of lights.