Bill S5335A version A - Patricia Canzoneri-Fitzpatrick
BILL NUMBER: S5335A SPONSOR: CANZONERI-FITZPATRICK TITLE OF BILL: An act to amend the criminal procedure law, in relation to setting bail for defendants that pose a threat to public safety PURPOSE: To allow greater judicial discretion in setting a securing order with respect to the dangerousness of a defendant. SUMMARY OF PROVISIONS: Section 1: Amends subdivision 1 of section 510.