Bill S9947A version A - Julia Salazar
BILL NUMBER: S9947A SPONSOR: SALAZAR TITLE OF BILL: An act to amend the criminal procedure law and the executive law, in relation to establishing an office of pretrial services PURPOSE OR GENERAL IDEA OF BILL: The purpose of this legislation is to improve the quality of pretrial services in New York. SUMMARY OF PROVISIONS: Section 1: Amends subdivisions 1, 2, 4, 5 and adds new 1-a, 2-a and 6 to section 510.